Added BootBox and feed DataTables with customer API. Added MembershipType to customer API

+++ CinemaWatch/Dtos/MovieDto.cs
@@ -15,7 +15,7 @@ namespace CinemaWatch.Dtos
         public string Name { get; set; }
 
         [Required]
-        public byte GenreId { get; set; }
+        public int GenreId { get; set; }
 
         public DateTime DateAdded { get; set; }
 
@@ -23,5 +23,7 @@ namespace CinemaWatch.Dtos
 
         [Range(1, 20)]
         public byte NumberInStock { get; set; }
+
+
     }
 }
